# project/doc/OpenAPI/Java(Spring)/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost
    description: Generated server url
...
paths:
  /user:
    get:
      tags:
        - UserController
      summary: 检测用户名是否被占用
      operationId: exist
      parameters:
        - name: username
          in: query
          description: 需要检测的用户名
          required: true
          schema:
            type: string
      responses:
        "4001":
          description: 用户名已被占用
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
        "4000":
          description: 用户名可用
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
    post:
      tags:
        - UserController
      summary: 用户注册
      description: 用户名不能重复
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "6001":
          description: 注册失败
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
        "1001":
          description: 验证码错误
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
        "6000":
          description: 注册成功
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
  /user/login:
    post:
      tags:
        - UserController
      summary: 用户登陆请求
      operationId: home
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "5000":
          description: 登陆成功
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
        "1001":
          description: 验证码错误
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
        "5001":
          description: 登陆失败
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
  /server:
    post:
      tags:
        - ServerController
      summary: 主业务接口，返回配乐文件的哈希值
      operationId: mainServer
      parameters:
        - name: map
          in: query
          required: true
          schema:
            type: object
            additionalProperties:
              type: object
      requestBody:
        content:
          application/json:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        default:
          description: 标准返回格式，data为生成文件的md5值
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
...
  /download/{md5}:
    get:
      tags:
        - DownloadController
      summary: 根据md5下载对应文件
      operationId: downFile
      parameters:
        - name: md5
          in: path
          description: 想下载文件的哈希值
          required: true
          schema:
            type: string
      responses:
        default:
          description: 返回文件的二进制数据
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
  /download/person/{index}:
    get:
      tags:
        - DownloadController
      summary: 下载历史数据文件
      description: 根据索引下载用户自身的历史数据
      operationId: downPersonalFile
      parameters:
        - name: origin
          in: query
          description: 若为1则返回原始文件，为0返回配乐后文件
          required: true
          schema:
            type: integer
            format: int32
        - name: index
          in: path
          description: 下载的数据的索引
          required: true
          schema:
            type: integer
            format: int32
      responses:
        default:
          description: 返回文件的二进制数据
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Result'
...
components:
  schemas:
    User:
      required:
        - captcha
        - password
        - username
      type: object
      properties:
        id:
          type: integer
          description: id值
          format: int32
        username:
          pattern: "^[a-zA-Z0-9_\\u4e00-\\u9fa5]+$"
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
        captcha:
          type: string
          description: 验证码，用于发送请求时验证
      description: 要登录的用户信息
    Result:
      type: object
      properties:
        code:
          type: integer
          description: 自定义响应代码
          format: int32
        data:
          type: object
          description: 传入前端的数据
          example: 二进制数据
        msg:
          type: string
          description: 传入前端的信息
      description: 通用返回结果
...